.history/requirements_20250716155843.txt
# SafeServe AI - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Machine Learning & Data Science
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
pyod==1.1.3
joblib==1.3.2

# Deep Learning & NLP
torch==2.1.0
transformers==4.35.2
accelerate==0.24.1

# API & HTTP
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0

# Environment & Configuration
python-dotenv==1.0.0

# Data Visualization
plotly==5.17.0

# Networking & Tunneling
pyngrok==6.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Audio/TTS (for voice features)
# TTS==0.20.2

# Optional: Additional ML models
# sentence-transformers==2.2.2

# System utilities
nest-asyncio==1.5.8
python-multipart==0.0.6

# Excel support
openpyxl==3.1.2

# Async support
asyncio==3.4.3